+++ src/test.cpp
@@ -1,3 +1,4 @@
+#include <cmath>
 #include <iostream>
 #include <numbers>
 
@@ -22,6 +23,16 @@ static int checks_passed = 0;
       std::cout << "CHECK_FALSE at line " << __LINE__ << " failed!\n"; \
     } else {                                                           \
       ++checks_passed;                                                 \
+    }                                                                  \
+  } while (false);
+
+#define CHECK_FLOAT(var, value)                                        \
+  do {                                                                 \
+    ++checks_checked;                                                  \
+    if ((var) > (value)-0.0001F && (var) < (value) + 0.0001F) {        \
+      ++checks_passed;                                                 \
+    } else {                                                           \
+      std::cout << "CHECK_FLOAT at line " << __LINE__ << " failed!\n"; \
     }                                                                  \
   } while (false);
 
@@ -85,8 +96,10 @@ int main() {
 
   // Test Separating_Axis_Collision check.
   {
-    SC_SACD_Generic_Box a{0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F};
-    SC_SACD_Generic_Box b{0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, 0.0F, 0.0F, 0.0F};
+    SC_SACD_Generic_Box a{
+        0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, SC_SACD_Mat4_Identity()};
+    SC_SACD_Generic_Box b{
+        0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F, SC_SACD_Mat4_Identity()};
 
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
 
@@ -95,44 +108,468 @@ int main() {
     b.x = -1.1F;
     CHECK_FALSE(SC_SACD_Generic_Box_Collision(&a, &b));
 
-    a.z_radians = std::numbers::pi_v<float> / 4.0F;
-    b.z_radians = a.z_radians;
+    a.transform = SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 4.0F);
+    b.transform = a.transform;
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
     b.x = 1.1F;
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
 
     b.x = 0.0F;
     b.y = 1.1F;
-    a.z_radians = 0.0F;
-    b.z_radians = 0.0F;
+    a.transform = SC_SACD_Mat4_Identity();
+    b.transform = a.transform;
     CHECK_FALSE(SC_SACD_Generic_Box_Collision(&a, &b));
     b.y = -1.1F;
     CHECK_FALSE(SC_SACD_Generic_Box_Collision(&a, &b));
 
-    a.z_radians = std::numbers::pi_v<float> / 4.0F;
-    b.z_radians = a.z_radians;
+    a.transform = SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 4.0F);
+    b.transform = a.transform;
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
     b.y = 1.1F;
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
 
     b.y = 0.0F;
-    a.z_radians = 0.0F;
-    b.z_radians = 0.0F;
+    a.transform = SC_SACD_Mat4_Identity();
+    b.transform = a.transform;
 
     b.z = 1.1F;
     CHECK_FALSE(SC_SACD_Generic_Box_Collision(&a, &b));
     b.z = -1.1F;
     CHECK_FALSE(SC_SACD_Generic_Box_Collision(&a, &b));
 
-    a.y_radians = std::numbers::pi_v<float> / 4.0F;
-    b.y_radians = a.y_radians;
+    a.transform = SC_SACD_Rotation_Mat4_YAxis(std::numbers::pi_v<float> / 4.0F);
+    b.transform = a.transform;
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
     b.z = 1.1F;
     CHECK_TRUE(SC_SACD_Generic_Box_Collision(&a, &b));
   }
 
+  // Test Sphere/Sphere collision check.
+  {
+    SC_SACD_Sphere a{0.0F, 0.0F, 0.0F, 1.0F};
+    SC_SACD_Sphere b{0.0F, 0.0F, 0.0F, 1.0F};
+    CHECK_TRUE(SC_SACD_Sphere_Collision(&a, &b));
+
+    a.x = 10.0F;
+    a.y = 10.0F;
+    a.z = 10.0F;
+    b.x = 10.0F;
+    b.y = 10.5F;
+    b.z = 10.0F;
+    CHECK_TRUE(SC_SACD_Sphere_Collision(&a, &b));
+    b.y = 12.0F;
+    b.z = 12.0F;
+    CHECK_FALSE(SC_SACD_Sphere_Collision(&a, &b));
+    b.y = 11.0F;
+    b.z = 11.0F;
+    CHECK_TRUE(SC_SACD_Sphere_Collision(&a, &b));
+  }
+
+  // Test Sphere/AABB collision check.
+  {
+    SC_SACD_Sphere sphere{0.0F, 0.0F, 0.0F, 1.0F};
+    SC_SACD_AABB_Box box{0.0F, 0.0F, 0.0F, 2.0F, 2.0F, 2.0F};
+
+    CHECK_TRUE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = -2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 0.0F;
+    sphere.y = 2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.y = -2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.y = 0.0F;
+    sphere.z = 2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.z = -2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.z = 0.0F;
+    sphere.x = 1.0F;
+    sphere.y = 1.0F;
+    CHECK_TRUE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = -1.0F;
+    sphere.y = -1.0F;
+    CHECK_TRUE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 2.0F;
+    sphere.y = 2.0F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = -2.0F;
+    sphere.y = -2.0F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 1.0F;
+    sphere.y = 0.0F;
+    sphere.z = 1.0F;
+    CHECK_TRUE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = -1.0F;
+    sphere.z = -1.0F;
+    CHECK_TRUE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 2.0F;
+    sphere.z = 2.0F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = -2.0F;
+    sphere.z = -2.0F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 1.0F;
+    sphere.z = 1.5F;
+    CHECK_TRUE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+
+    sphere.x = 1.0F;
+    sphere.z = 2.0F;
+    CHECK_FALSE(SC_SACD_Sphere_AABB_Box_Collision(&sphere, &box));
+  }
+
+  // Test Sphere/Generic_Box collision check.
+  {
+    SC_SACD_Sphere sphere{0.0F, 0.0F, 0.0F, 1.0F};
+    SC_SACD_Generic_Box box{
+        0.0F,
+        0.0F,
+        0.0F,
+        2.0F,
+        2.0F,
+        2.0F,
+        SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 4.0F)};
+
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = -2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = -2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 0.0F;
+    sphere.y = 2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.y = -2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.y = 2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.y = -2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.y = 0.0F;
+    sphere.z = 2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.z = -2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.z = 1.9F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.z = -1.9F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 1.5F;
+    sphere.y = 1.5F;
+    sphere.z = 0.0F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 1.4F;
+    sphere.y = 1.4F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 2.2F;
+    sphere.y = 0.7929F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 2.1F;
+    sphere.y = 0.6929F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = -1.5F;
+    sphere.y = -1.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = -1.4F;
+    sphere.y = -1.4F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = -2.2F;
+    sphere.y = -0.7929F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = -2.1F;
+    sphere.y = -0.6929F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    box.x = 10.0F;
+    box.y = -10.0F;
+    box.z = 13.0F;
+
+    sphere.y = -10.0F;
+    sphere.z = 13.0F;
+    sphere.x = 10.0F + 2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 10.0F + -2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 10.0F + 2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 10.0F + -2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + 0.0F;
+    sphere.y = -10.0F + 2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.y = -10.0F + -2.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.y = -10.0F + 2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.y = -10.0F + -2.3F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.y = -10.0F + 0.0F;
+    sphere.z = 13.0F + 2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.z = 13.0F + -2.1F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.z = 13.0F + 1.9F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.z = 13.0F + -1.9F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + 1.5F;
+    sphere.y = -10.0F + 1.5F;
+    sphere.z = 13.0F + 0.0F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 10.0F + 1.4F;
+    sphere.y = -10.0F + 1.4F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + 2.2F;
+    sphere.y = -10.0F + 0.7929F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + 2.1F;
+    sphere.y = -10.0F + 0.6929F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + -1.5F;
+    sphere.y = -10.0F + -1.5F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+    sphere.x = 10.0F + -1.4F;
+    sphere.y = -10.0F + -1.4F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + -2.2F;
+    sphere.y = -10.0F + -0.7929F;
+    CHECK_FALSE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+
+    sphere.x = 10.0F + -2.1F;
+    sphere.y = -10.0F + -0.6929F;
+    CHECK_TRUE(SC_SACD_Sphere_Box_Collision(&sphere, &box));
+  }
+
+  // Test matrix/vector multiplication.
+  {
+    SC_SACD_Mat4 mat_a{1.0F, 2.0F,  3.0F,  4.0F,  5.0F,  6.0F,  7.0F,  8.0F,
+                       9.0F, 10.0F, 11.0F, 12.0F, 13.0F, 14.0F, 15.0F, 16.0F};
+
+    SC_SACD_Mat4 mat_b = SC_SACD_Mat4_Identity();
+
+    {
+      auto result = SC_SACD_Mat4_Mult(&mat_a, &mat_b);
+      CHECK_TRUE(mat_a.x0 == result.x0);
+      CHECK_TRUE(mat_a.x1 == result.x1);
+      CHECK_TRUE(mat_a.x2 == result.x2);
+      CHECK_TRUE(mat_a.x3 == result.x3);
+      CHECK_TRUE(mat_a.y0 == result.y0);
+      CHECK_TRUE(mat_a.y1 == result.y1);
+      CHECK_TRUE(mat_a.y2 == result.y2);
+      CHECK_TRUE(mat_a.y3 == result.y3);
+      CHECK_TRUE(mat_a.z0 == result.z0);
+      CHECK_TRUE(mat_a.z1 == result.z1);
+      CHECK_TRUE(mat_a.z2 == result.z2);
+      CHECK_TRUE(mat_a.z3 == result.z3);
+      CHECK_TRUE(mat_a.w0 == result.w0);
+      CHECK_TRUE(mat_a.w1 == result.w1);
+      CHECK_TRUE(mat_a.w2 == result.w2);
+      CHECK_TRUE(mat_a.w3 == result.w3);
+    }
+
+    mat_b.x0 = 2.0F;
+    mat_b.y1 = 0.0F;
+    mat_b.z2 = 0.0F;
+    {
+      auto result = SC_SACD_Mat4_Mult(&mat_a, &mat_b);
+      CHECK_FLOAT(result.x0, 2.0F);
+      CHECK_FLOAT(result.y0, 10.0F);
+      CHECK_FLOAT(result.z0, 18.0F);
+      CHECK_FLOAT(result.w0, 26.0F);
+      CHECK_FLOAT(result.x1, 0.0F);
+      CHECK_FLOAT(result.y1, 0.0F);
+      CHECK_FLOAT(result.z1, 0.0F);
+      CHECK_FLOAT(result.w1, 0.0F);
+      CHECK_FLOAT(result.x2, 0.0F);
+      CHECK_FLOAT(result.y2, 0.0F);
+      CHECK_FLOAT(result.z2, 0.0F);
+      CHECK_FLOAT(result.w2, 0.0F);
+      CHECK_FLOAT(result.x3, 4.0F);
+      CHECK_FLOAT(result.y3, 8.0F);
+      CHECK_FLOAT(result.z3, 12.0F);
+      CHECK_FLOAT(result.w3, 16.0F);
+    }
+
+    mat_b = SC_SACD_Mat4_Identity();
+    SC_SACD_Vec3 vec_a{1.0F, 0.0F, 0.0F};
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_b, vec_a);
+      CHECK_TRUE(result.x == vec_a.x);
+      CHECK_TRUE(result.y == vec_a.y);
+      CHECK_TRUE(result.z == vec_a.z);
+    }
+
+    // Rotations about each axis.
+    mat_a = SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 2.0F);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 0.0001F && result.x > -0.0001F);
+      CHECK_TRUE(result.y < 1.0001F && result.y > 0.9999F);
+      CHECK_TRUE(result.z < 0.0001F && result.z > -0.0001F);
+    }
+
+    mat_a = SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float>);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < -0.9999F && result.x > -1.0001F);
+      CHECK_TRUE(result.y < 0.0001F && result.y > -0.0001F);
+      CHECK_TRUE(result.z < 0.0001F && result.z > -0.0001F);
+    }
+
+    mat_a =
+        SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> * 3.0F / 2.0F);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 0.0001F && result.x > -0.0001F);
+      CHECK_TRUE(result.y < -0.9999F && result.y > -1.0001F);
+      CHECK_TRUE(result.z < 0.0001F && result.z > -0.0001F);
+    }
+
+    mat_a = SC_SACD_Rotation_Mat4_XAxis(std::numbers::pi_v<float> / 2.0F);
+    vec_a.x = 0.0F;
+    vec_a.y = 1.0F;
+    vec_a.z = 0.0F;
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 0.0001F && result.x > -0.0001F);
+      CHECK_TRUE(result.y < 0.0001F && result.y > -0.0001F);
+      CHECK_TRUE(result.z < 1.0001F && result.z > 0.9999F);
+    }
+
+    mat_a = SC_SACD_Rotation_Mat4_XAxis(std::numbers::pi_v<float>);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 0.0001F && result.x > -0.0001F);
+      CHECK_TRUE(result.y < -0.9999F && result.y > -1.0001F);
+      CHECK_TRUE(result.z < 0.0001F && result.z > -0.0001F);
+    }
+
+    mat_a =
+        SC_SACD_Rotation_Mat4_XAxis(std::numbers::pi_v<float> * 3.0F / 2.0F);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 0.0001F && result.x > -0.0001F);
+      CHECK_TRUE(result.y < 0.0001F && result.y > -0.0001F);
+      CHECK_TRUE(result.z < -0.9999F && result.z > -1.0001F);
+    }
+
+    mat_a = SC_SACD_Rotation_Mat4_YAxis(std::numbers::pi_v<float> / 2.0F);
+    vec_a.x = 0.0F;
+    vec_a.y = 0.0F;
+    vec_a.z = 1.0F;
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 1.0001F && result.x > 0.9999F);
+      CHECK_TRUE(result.y < 0.0001F && result.y > -0.0001F);
+      CHECK_TRUE(result.z < 0.0001F && result.z > -0.0001F);
+    }
+
+    mat_a = SC_SACD_Rotation_Mat4_YAxis(std::numbers::pi_v<float>);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < 0.0001F && result.x > -0.0001F);
+      CHECK_TRUE(result.y < 0.0001F && result.y > -0.0001F);
+      CHECK_TRUE(result.z < -0.9999F && result.z > -1.0001F);
+    }
+
+    mat_a =
+        SC_SACD_Rotation_Mat4_YAxis(std::numbers::pi_v<float> * 3.0F / 2.0F);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_TRUE(result.x < -0.9999F && result.x > -1.0001F);
+      CHECK_TRUE(result.y < 0.0001F && result.y > -0.0001F);
+      CHECK_TRUE(result.z < 0.0001F && result.z > -0.0001F);
+    }
+
+    // Combined axis rotation.
+    vec_a.x = 1.0F;
+    vec_a.y = 0.0F;
+    vec_a.z = 0.0F;
+    mat_a = SC_SACD_Rotation_Mat4_YAxis(std::numbers::pi_v<float> / 4.0F);
+    mat_b = SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 4.0F);
+    // Apply mat_a, then mat_b.
+    mat_a = SC_SACD_Mat4_Mult(&mat_b, &mat_a);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_FLOAT(result.x, 0.5F);
+      CHECK_FLOAT(result.y, 0.5F);
+      CHECK_FLOAT(result.z, -std::sqrt(2.0F) / 2.0F);
+    }
+    // Apply another rotation on combined mat_a.
+    mat_b = SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 4.0F);
+    mat_a = SC_SACD_Mat4_Mult(&mat_b, &mat_a);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_FLOAT(result.x, 0.0F);
+      CHECK_FLOAT(result.y, std::sqrt(2.0F) / 2.0F);
+      CHECK_FLOAT(result.z, -std::sqrt(2.0F) / 2.0F);
+    }
+    // Apply another rotation on combined mat_a.
+    mat_b = SC_SACD_Rotation_Mat4_XAxis(std::numbers::pi_v<float> / 2.0F);
+    mat_a = SC_SACD_Mat4_Mult(&mat_b, &mat_a);
+    {
+      auto result = SC_SACD_Mat4_Vec3_Mult(&mat_a, vec_a);
+      CHECK_FLOAT(result.x, 0.0F);
+      CHECK_FLOAT(result.y, std::sqrt(2.0F) / 2.0F);
+      CHECK_FLOAT(result.z, std::sqrt(2.0F) / 2.0F);
+    }
+  }
+
+  // Mat4 rotation and translation.
+  {
+    SC_SACD_Mat4 mat_a = SC_SACD_Translate_Mat4(1.0F, 1.0F, 1.0F);
+    SC_SACD_Mat4 mat_b =
+        SC_SACD_Rotation_Mat4_ZAxis(std::numbers::pi_v<float> / 4.0F);
+    mat_a = SC_SACD_Mat4_Mult(&mat_b, &mat_a);
+    mat_b = SC_SACD_Translate_Mat4(0.0F, 0.0F, -1.0F);
+    mat_a = SC_SACD_Mat4_Mult(&mat_b, &mat_a);
+
+    {
+      auto result =
+          SC_SACD_Mat4_Vec3_Mult(&mat_a, SC_SACD_Vec3{0.0F, 0.0F, 0.0F});
+      CHECK_FLOAT(result.x, 0.0F);
+      CHECK_FLOAT(result.z, 0.0F);
+      CHECK_FLOAT(result.y, std::sqrt(2.0F));
+    }
+  }
   std::cout << "Checks checked: " << checks_checked << '\n'
             << "Checks passed:  " << checks_passed << '\n';
 
-  return 0;
+  return checks_checked == checks_passed ? 0 : 1;
 }