+++ Changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 2.0.0
+
+Add collision detection with Spheres.
+
+Add some more vector/matrix math stuff.
+
+Replace Mat3 with Mat4.
+
+Replace xyz-axis rotation floats in GenericBox with Mat4.
+
 ## Version 1.0.0
 
 First version of this library.